QA finish

+++ laravel/app/Answer.php
@@ -57,22 +57,32 @@ class Answer extends Model
 
     public function read()
     {
-        if(!rq('id')&&!rq('question_id'))
+        if(!rq('id')&&!rq('question_id')&&!rq('user_id'))
         {
             return ['status'=>0,'msg'=>'need id or question_id'];
         }
         if(rq('id'))
         {
-            $answer = $this->find(rq('id'));
+            $answer = $this::with('user','question','users','comments')->find(rq('id'));
+            $answer->commentCount = count($answer->comments);
             if(!$answer)
                 return ['status'=>0,'msg'=>' answer not exists'];
             return ['status'=>1,'data'=>$answer];
         }
 
+        if(rq('user_id'))
+        {
+            $user = user_ins()->find(rq('user_id'));
+            if(!$user)
+                return err('user not exists');
+            $answer = $this::with('user','question','users','comments')->where('user_id',rq('user_id'))->orderby('created_at')->get();
+            return suc($answer->toArray());
+        }
+
         $question = question_ins()->find(rq('question_id'));
         if(!$question)
             return ['status'=>0,'msg'=>'question not exists'];
-        $answers = $this->where(['question_id'=>rq('question_id')])->orderby('created_at')->get()->keyby('id');
+        $answers = $this->with('user','users')->where(['question_id'=>rq('question_id')])->orderby('created_at')->get()->keyby('id');
         return ['status'=>1,'data'=>$answers];
     }
 
@@ -92,15 +102,23 @@ class Answer extends Model
         if(!$answer)
             return ['status'=>0,'msg'=>'answer not exists'];
         $answerUser = $answer->users()->where(['user_id'=>session('user_id')])->first();
-
+        if(session()->get('user_id') == $answer->user_id)
+        {
+            return err('cannot vote yourself');
+        }
        if($answerUser)
        {
+            $vote_o = $answerUser->pivot->vote;
            $answerUser->pivot->delete();
+           if($vote_o == $vote)
+           {
+               return suc();
+           }
        }
 
         $answer->users()->attach(session('user_id'),['vote' => $vote]);
 
-        return ['status'=>0 ] ;
+        return ['status'=>1 ] ;
 
         //dd(DB::getQueryLog());
 
@@ -110,11 +128,26 @@ class Answer extends Model
 
     }
 
+    public function user()
+    {
+        return $this->belongsTo('\App\User')->select(array('id','username','intro'));
+    }
+
     public function users()
     {
-        return $this->belongsToMany('\App\User')
+        return $this->belongsToMany('\App\User')->select(array('users.id','users.username'))
             ->withPivot('vote')
             ->withTimestamps()
             ;
+    }
+
+    public function question()
+    {
+        return $this->belongsTo('\App\Question');
+    }
+
+    public function comments()
+    {
+        return $this->hasMany('\App\Comment');
     }
 }
